requirements.txt
pandas
boto3
python-dotenv
requests
flask
flask-restful
fsspec
s3fs